+++ lib/rrj/tools/gem/log.rb
@@ -23,6 +23,9 @@ module RubyRabbitmqJanus
         UNKNOWN: Logger::UNKNOWN
       }.freeze
 
+      # Sensitive data
+      SENSITIVES = %i(admin_secret apisecret).freeze
+
       attr_reader :level
 
       # Returns a new instance to Log and use `Tag` element for each line
@@ -43,41 +46,41 @@ module RubyRabbitmqJanus
       #
       # @param message [String] Message writing in warning level in log
       def unknown(message)
-        write_tag { @logs.unknown(message) }
+        write_tag { @logs.unknown(filter(message)) }
       end
 
       # Write a message in log with a `FATAL` level
       #
       # @param message [String] Message writing in warning level in log
       def fatal(message)
-        write_tag { @logs.fatal(message) } if test_level?(Logger::FATAL)
+        write_tag { @logs.fatal(filter(message)) } if test_level?(Logger::FATAL)
       end
 
       # Write a message in log with a `ERROR` level
       #
       # @param message [String] Message writing in warning level in log
       def error(message)
-        write_tag { @logs.error(message) } if test_level?(Logger::ERROR)
+        write_tag { @logs.error(filter(message)) } if test_level?(Logger::ERROR)
       end
 
       # Write a message in log with a `WARN` level
       # @param message [String] Message writing in warning level in log
       def warn(message)
-        write_tag { @logs.warn(message) } if test_level?(Logger::WARN)
+        write_tag { @logs.warn(filter(message)) } if test_level?(Logger::WARN)
       end
 
       # Write a message in log with a `INFO` level
       #
       # @param message [String] Message writing in info level in log
       def info(message)
-        write_tag { @logs.info(message) } if test_level?(Logger::INFO)
+        write_tag { @logs.info(filter(message)) } if test_level?(Logger::INFO)
       end
 
       # Write a message in log with a `DEBUG` level
       #
       # @param message [String] Message writing in debug level in log
       def debug(message)
-        write_tag { @logs.debug(message) } if test_level?(Logger::DEBUG)
+        write_tag { @logs.debug(filter(message)) } if test_level?(Logger::DEBUG)
       end
 
       # @return [RubyRabbitmqJanus::Tools::Log] the instance to logger
@@ -118,6 +121,18 @@ module RubyRabbitmqJanus
       def write_tag
         @logs.tagged(@logs.progname) { yield }
       end
+
+      # @todo fix replace
+      def filter_sensitive_data(log)
+        msg = log
+        SENSITIVES.each do |word|
+          msg = log.gsub(/\"#{word}\".*\".*\"/, "\"#{word}\":\"[FILTERED]\"") \
+            if log.include?(word.to_s)
+        end
+        msg
+      end
+
+      alias filter filter_sensitive_data
     end
   end
 end